refactor(ui/web): migrate LeverageSlider from @starboard
